Adds compatibility with latest onegov.core release

+++ src/onegov.town/onegov/town/models/resource.py
@@ -25,24 +25,22 @@ class SharedMethods(object):
     def remove_expired_reservation_sessions(self, libres_context,
                                             expiration_date=None):
         session = libres_context.get_service('session_provider').session()
-        scheduler = self.get_scheduler(libres_context)
+        queries = self.get_scheduler(libres_context).queries
 
-        find = scheduler.queries.find_expired_reservation_sessions
-        remove = scheduler.queries.remove_expired_reservation_sessions
+        expired_sessions = queries.find_expired_reservation_sessions(
+            expiration_date)
 
-        expired_sessions = find(expiration_date)
+        if expired_sessions:
+            query = session.query(Reservation).with_entities(Reservation.token)
+            query = query.filter(Reservation.session_id.in_(expired_sessions))
+            tokens = set(r[0] for r in query.all())
 
-        query = session.query(Reservation).with_entities(Reservation.token)
-        query = query.filter(Reservation.session_id.in_(expired_sessions))
-        tokens = set(r[0] for r in query.all())
+            query = session.query(FormSubmission)
+            query = query.filter(FormSubmission.name == None)
+            query = query.filter(FormSubmission.id.in_(tokens))
 
-        remove(expiration_date)
-
-        query = session.query(FormSubmission)
-        query = query.filter(FormSubmission.name == None)
-        query = query.filter(FormSubmission.id.in_(tokens))
-
-        query.delete('fetch')
+            query.delete('fetch')
+            queries.remove_expired_reservation_sessions(expiration_date)
 
 
 class SearchableResource(ORMSearchable):
